+++ src/event-reduce.ts
@@ -16,7 +16,6 @@ import type {
     StringKeys,
     RxDocumentData
 } from './types';
-import { runPluginHooks } from './hooks';
 import { rxChangeEventToEventReduceChangeEvent } from './rx-change-event';
 import { clone, ensureNotFalsy } from './util';
 import { normalizeMangoQuery } from './rx-query-helper';
@@ -74,7 +73,6 @@ export function getQueryParams<RxDocType>(
                 docB,
                 rxQuery
             };
-            runPluginHooks('preSortComparator', sortComparatorData);
             return sortComparator(sortComparatorData.docA, sortComparatorData.docB);
         };
 
@@ -92,8 +90,6 @@ export function getQueryParams<RxDocType>(
                 doc,
                 rxQuery
             };
-            runPluginHooks('preQueryMatcher', queryMatcherData);
-
             return queryMatcher(queryMatcherData.doc);
         };