+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - REFACTORED the encryption plugin, it is no longer a plugin but now a wrapper around any other RxStorage.
   - It allows to run the encryption inside of a [Worker RxStorage](./rx-storage-worker.md) instead of running it in the main JavaScript process.
   - It allows do use asynchronous crypto function like [WebCrypto](https://developer.mozilla.org/en-US/docs/Web/API/Web_Crypto_API)
+- REMOVED many unused plugin hooks because they decreased the performance.
 
 <!-- ADD new changes here! -->